Windows API을 사용한 채팅, 서버 - 클라이언트 채팅 패킷 통신, 채팅 개행

+++ Client/GameCoding/Creature.h
@@ -20,12 +20,12 @@ public:
 	virtual void OnDamaged(Creature* attacker);
 	
 	void KnockBack();
-	float SetWait(float time) { return _wait = GetTickCount64() + time; }
+	uint64 SetWait(uint64 time) { return _wait = GetTickCount64() + time; }
 
 	Protocol::OBJECT_TYPE GetType() { return this->info.objecttype(); }
 
 protected:
-	float _wait = 0;
-	float _now = 0;
+	uint64 _wait = 0;
+	uint64 _now = 0;
 };
